main és game osztály elválasztása
